New tools to support the creation of mil feat geodatabase template
Two GP tools added to a new toolbox.
CreateTemplateGDB reads a pair of CSV files to create the core of a new
geodatabase.  It then calls the  AddFieldsFromSchema tool to add
specified fields to each feature class in that geodatabase.

Note: there appears to be a bug in the latest 1.1.1 update of Pro.
Toolboxes created in the Project tab cannot be renamed with spaces in
the Label.  One would expect that the Name of the toolbox should be
spaceless and become the name of the tbx file.  One would expect the
Label property to be the property that drives how the toolbox is listed
in the Project tab, and therefore allow spaces.  This no longer appears
to be the case.