Update old link to lines-areas-labels.stylx
Chris, the link to the 'lines-areas-labels-base-template.stylx' navigated to a "retired-dont-use" section of the repo, with a version of the .stylx from November 2014. You can see that I changed the path of this folder to an updated version of the template stylx, located in "core data/stylx files." Later on in the readme (line 23) the proper location of 'lines-areas-labels-base-template.stylx' is mentioned anyway.