Military Features 2525D Reogranization
This reorganization of the 2525D section of the Military Features Data
Repository includes updated versions of the MilitaryOverlay GDB, a new
"overlay_templates" folder with updated templates, and a reorganization
of the "sample_data" folder into a new "test_data" folder. In addition,
this commit includes a fix to issue #179, removing the SVG Dash
Workaround folder.