Updating Complete Test Datasets + Scripts and Models for C_Bc2
These complete test datasets show every symbol for each appendix that
can be rendered with our legacy renderer using the C or Bc2 setting. In
addition, included are a script and a toolbox. The python script allows
for users to filter out source CSVs to get specific appendices, whereas
the models in the toolbox create the test data and match it up to the
specific CSVs, giving an editable test dataset as an output. In
addition, included in this PR is a list of control measures that don't
follow the point ordering dictated by 2525D, in order to increase
transparency for users of the Military Overlay template.